V.3 Logic UPdate
This includes logic that prevents the pi from starting with the relays all energized. (not good for hvac)
As well as some work on presenting data more dependably. The logic is not quite complete, but the framework is there.

(TO ADD NEXT)
Figure out why the Relay doesn't go to the off state before switching between heat and cool. (thats not good to have happen) It should have a 10-20 second delay before switching between them.

Tracking system states and time stamping them to be sent to the database.

Better UI elements. It's not pretty. But it's still pretty cool. 😎