V 1.0
A complete overhaul and rewrite of the main Thermostat logic. It monitors the temp while in states to determine where to go next. Once a state is completed, it runs into a between states mode and starts a timer. This is to ensure your heat pump/compressor has time to rest. These are all configurable in the config file as well as the plethora of variables that determine how cool or warm to make it.

It is also all configurable from the front end. Touch friendly UI is mobile friendly as well as larger screen friendly.