V1.4
Bug fixes.
At first the thermostat was NOT reporting the state change to between states. After my initial fix it was causing the SQL server to crash. SO I fixed that too.

UPDATES.
Prepping a few things for app integration. This will support V1 of the flutter app that will be available in its own repo.

Some of those updates include proper naming and ip capturing and whatnot for the server to send info correctly to the app during initial setup.